Dodanie klienta na Androida

+++ AirHockey/tescixyODE/cSerwer.h
@@ -4,9 +4,9 @@
 
 struct sRamkaV
 {
-int xV;
-int yV;
-int zV;
+float xV;
+float yV;
+float zV;
 };
